+++ sim/game/cards/dm12/angel_command.go
@@ -1,0 +1,22 @@
+package dm12
+
+import (
+	"duel-masters/game/civ"
+	"duel-masters/game/family"
+	"duel-masters/game/fx"
+	"duel-masters/game/match"
+)
+
+// ValkyerStarstormElemental ...
+func ValkyerStarstormElemental(c *match.Card) {
+
+	c.Name = "Valkyer, Starstorm Elemental"
+	c.Power = 7000
+	c.Civs = []string{civ.Light}
+	c.Family = []string{family.AngelCommand}
+	c.ManaCost = 5
+	c.ManaRequirement = []string{civ.Light}
+
+	c.Use(fx.Creature, fx.Blocker(), fx.CantAttackPlayers)
+
+}